Move query utils from `aiida.cmdline` to `aiida.tools`
These utilities are useful not only for the CLI but also in user
applications, such as notebooks or in plugins. It makes more sense to
have these in `aiida.tools`. The original locations import the new
modules to keep backwards compatibility and a deprecation warning is
added to the `aiida.cmdline` modules.

The choices and default of the `--project` option of the `process list`
had to be made indirect because of the move. The reason is that they
rely on the `CalculationQueryBuilder` class, which now has moved to the
`aiida.tools` module. This means that it should be imported on the top
level of the `aiida.cmdline`, however, this is not permissible as the
load time of this is significant and it will slow down tab-completion,
which will have to load the cmdline module. The work around is to make
the import of the class indirect by using the `LazyChoice` class as the
type of the option and by making the default a lambda that calls a
function that performs the actual import.
